v8.3

+++ customize.sh
@@ -104,7 +104,7 @@ ui_print "$FILE"
 ui_print "  Please wait..."
 if ! grep -q $NAME $FILE; then
   ui_print "  Function not found."
-  ui_print "  Using new $DIR$LIB"
+  ui_print "  Using new $DIR/$LIB"
   mv -f $MODPATH/system_support$DIR/$LIB $MODPATH/system$DIR
 fi
 ui_print " "
@@ -219,7 +219,12 @@ fi
 ui_print "- Cleaning..."
 if [ $DOLBY == true ]; then
   PKGS=`cat $MODPATH/package-dolby.txt`
-  rm -f /data/vendor/dolby/dax_sqlite3.db
+  if [ "`grep_prop dolby.mod $OPTIONALS`" == 0 ]; then
+    rm -f /data/vendor/dolby/dax_sqlite3.db
+  else
+    rm -f /data/vendor/dolby/dap_sqlite3.db
+    sed -i 's|dax_sqlite3.db|dap_sqlite3.db|g' $MODPATH/uninstall.sh
+  fi
 else
   PKGS=`cat $MODPATH/package.txt`
 fi
@@ -270,8 +275,12 @@ done
 
 # conflict
 if [ $DOLBY == true ]; then
-  NAMES="dolbyatmos DolbyAtmos DolbyAudio MotoDolby
-         dsplus Dolby"
+  if [ "`grep_prop dolby.mod $OPTIONALS`" == 0 ]; then
+    NAMES="dolbyatmos DolbyAtmos DolbyAudio MotoDolby
+           dsplus Dolby"
+  else
+    NAMES="dolbyatmos DolbyAtmos DolbyAudio MotoDolby"
+  fi
   conflict
   NAMES=MiSound
   FILE=/data/adb/modules/$NAMES/module.prop
@@ -397,47 +406,58 @@ if echo $MAGISK_VER | grep -Eq 'delta|Delta|kitsune'\
     MOUNT=`mount | grep $MAGISKTMP/preinit`
     BLOCK=`echo $MOUNT | sed 's| on.*||g'`
     DIR=`mount | sed "s|$MOUNT||g" | grep -m 1 $BLOCK`
-    EIMDIR=`echo $DIR | sed "s|$BLOCK on ||g" | sed 's| type.*||g'`/early-mount.d
-  elif ! $ISENCRYPTED; then
-    EIMDIR=/data/adb/early-mount.d
-  elif [ -d /data/unencrypted ]\
-  && ! grep ' /data ' /proc/mounts | grep -q dm-\
-  && grep ' /data ' /proc/mounts | grep -q ext4; then
-    EIMDIR=/data/unencrypted/early-mount.d
-  elif grep ' /cache ' /proc/mounts | grep -q ext4; then
-    EIMDIR=/cache/early-mount.d
-  elif grep ' /metadata ' /proc/mounts | grep -q ext4; then
-    EIMDIR=/metadata/early-mount.d
-  elif grep ' /persist ' /proc/mounts | grep -q ext4; then
-    EIMDIR=/persist/early-mount.d
-  elif grep ' /mnt/vendor/persist ' /proc/mounts | grep -q ext4; then
-    EIMDIR=/mnt/vendor/persist/early-mount.d
-  elif grep ' /cust ' /proc/mounts | grep -q ext4; then
-    EIMDIR=/cust/early-mount.d
-  elif [ "$MAGISK_VER_CODE" -ge 26000 ]\
-  && [ -d /data/unencrypted ]\
-  && ! grep ' /data ' /proc/mounts | grep -q dm-\
-  && grep ' /data ' /proc/mounts | grep -q f2fs; then
-    EIMDIR=/data/unencrypted/early-mount.d
-  elif [ "$MAGISK_VER_CODE" -ge 26000 ]\
-  && grep ' /cache ' /proc/mounts | grep -q f2fs; then
-    EIMDIR=/cache/early-mount.d
-  elif [ "$MAGISK_VER_CODE" -ge 26000 ]\
-  && grep ' /metadata ' /proc/mounts | grep -q f2fs; then
-    EIMDIR=/metadata/early-mount.d
-  elif [ "$MAGISK_VER_CODE" -ge 26000 ]\
-  && grep ' /persist ' /proc/mounts | grep -q f2fs; then
-    EIMDIR=/persist/early-mount.d
-  elif [ "$MAGISK_VER_CODE" -ge 26000 ]\
-  && grep ' /mnt/vendor/persist ' /proc/mounts | grep -q f2fs; then
-    EIMDIR=/mnt/vendor/persist/early-mount.d
-  elif [ "$MAGISK_VER_CODE" -ge 26000 ]\
-  && grep ' /cust ' /proc/mounts | grep -q f2fs; then
-    EIMDIR=/cust/early-mount.d
-  else
-    EIM=false
-    ui_print "- Unable to find early init mount directory"
-    ui_print " "
+    DIR=`echo $DIR | sed "s|$BLOCK on ||g" | sed 's| type.*||g'`
+    if [ "$DIR" ]; then
+      EIMDIR=$DIR/early-mount.d
+    else
+      ui_print "! It seems Magisk early init mount directory is not"
+      ui_print "  activated yet. Please reinstall Magisk.zip via Magisk app"
+      ui_print "  (not via Recovery)."
+      ui_print " "
+    fi
+  fi
+  if [ ! "$EIMDIR" ]; then
+    if ! $ISENCRYPTED; then
+      EIMDIR=/data/adb/early-mount.d
+    elif [ -d /data/unencrypted ]\
+    && ! grep ' /data ' /proc/mounts | grep -q dm-\
+    && grep ' /data ' /proc/mounts | grep -q ext4; then
+      EIMDIR=/data/unencrypted/early-mount.d
+    elif grep ' /cache ' /proc/mounts | grep -q ext4; then
+      EIMDIR=/cache/early-mount.d
+    elif grep ' /metadata ' /proc/mounts | grep -q ext4; then
+      EIMDIR=/metadata/early-mount.d
+    elif grep ' /persist ' /proc/mounts | grep -q ext4; then
+      EIMDIR=/persist/early-mount.d
+    elif grep ' /mnt/vendor/persist ' /proc/mounts | grep -q ext4; then
+      EIMDIR=/mnt/vendor/persist/early-mount.d
+    elif grep ' /cust ' /proc/mounts | grep -q ext4; then
+      EIMDIR=/cust/early-mount.d
+    elif [ "$MAGISK_VER_CODE" -ge 26000 ]\
+    && [ -d /data/unencrypted ]\
+    && ! grep ' /data ' /proc/mounts | grep -q dm-\
+    && grep ' /data ' /proc/mounts | grep -q f2fs; then
+      EIMDIR=/data/unencrypted/early-mount.d
+    elif [ "$MAGISK_VER_CODE" -ge 26000 ]\
+    && grep ' /cache ' /proc/mounts | grep -q f2fs; then
+      EIMDIR=/cache/early-mount.d
+    elif [ "$MAGISK_VER_CODE" -ge 26000 ]\
+    && grep ' /metadata ' /proc/mounts | grep -q f2fs; then
+      EIMDIR=/metadata/early-mount.d
+    elif [ "$MAGISK_VER_CODE" -ge 26000 ]\
+    && grep ' /persist ' /proc/mounts | grep -q f2fs; then
+      EIMDIR=/persist/early-mount.d
+    elif [ "$MAGISK_VER_CODE" -ge 26000 ]\
+    && grep ' /mnt/vendor/persist ' /proc/mounts | grep -q f2fs; then
+      EIMDIR=/mnt/vendor/persist/early-mount.d
+    elif [ "$MAGISK_VER_CODE" -ge 26000 ]\
+    && grep ' /cust ' /proc/mounts | grep -q f2fs; then
+      EIMDIR=/cust/early-mount.d
+    else
+      EIM=false
+      ui_print "- Unable to find early init mount directory"
+      ui_print " "
+    fi
   fi
   if [ -d ${EIMDIR%/early-mount.d} ]; then
     mkdir -p $EIMDIR
@@ -617,16 +637,19 @@ if [ $DOLBY == true ]; then
         $MAGISKTMP/mirror/*/*/etc/vintf/manifest/*.xml
         /*/etc/vintf/manifest/*.xml /*/*/etc/vintf/manifest/*.xml"
   if [ "`grep_prop dolby.skip.vendor $OPTIONALS`" != 1 ]\
+  && ! df -h $VENDOR | grep 100%\
   && ! grep -A2 vendor.dolby.hardware.dms $FILE | grep -q 1.0; then
     FILE=$VENDOR/etc/vintf/manifest.xml
     patch_manifest
   fi
   if [ "`grep_prop dolby.skip.system $OPTIONALS`" != 1 ]\
+  && ! df -h $SYSTEM | grep 100%\
   && ! grep -A2 vendor.dolby.hardware.dms $FILE | grep -q 1.0; then
     FILE=$SYSTEM/etc/vintf/manifest.xml
     patch_manifest
   fi
   if [ "`grep_prop dolby.skip.system_ext $OPTIONALS`" != 1 ]\
+  && ! df -h $SYSTEM_EXT | grep 100%\
   && ! grep -A2 vendor.dolby.hardware.dms $FILE | grep -q 1.0; then
     FILE=$SYSTEM_EXT/etc/vintf/manifest.xml
     patch_manifest
@@ -640,10 +663,6 @@ if [ $DOLBY == true ]; then
       ui_print "  You can fix this by using Magisk Delta/Kitsune Mask."
       ui_print " "
     fi
-    FILES="$MAGISKTMP/mirror/*/etc/vintf/manifest.xml
-           $MAGISKTMP/mirror/*/*/etc/vintf/manifest.xml
-           /*/etc/vintf/manifest.xml /*/*/etc/vintf/manifest.xml"
-    restore
   fi
 fi
 
@@ -654,17 +673,20 @@ if [ $DOLBY == true ]; then
         /*/etc/selinux/*_hwservice_contexts
         /*/*/etc/selinux/*_hwservice_contexts"
   if [ "`grep_prop dolby.skip.vendor $OPTIONALS`" != 1 ]\
+  && ! df -h $VENDOR | grep 100%\
   && ! grep -Eq 'u:object_r:hal_dms_hwservice:s0|u:object_r:default_android_hwservice:s0' $FILE; then
     FILE=$VENDOR/etc/selinux/vendor_hwservice_contexts
     patch_hwservice
   fi
   if [ "`grep_prop dolby.skip.system $OPTIONALS`" != 1 ]\
+  && ! df -h $SYSTEM | grep 100%\
   && ! grep -Eq 'u:object_r:hal_dms_hwservice:s0|u:object_r:default_android_hwservice:s0' $FILE; then
     FILE=$SYSTEM/etc/selinux/plat_hwservice_contexts
     patch_hwservice
   fi
   if [ "`grep_prop dolby.skip.system_ext $OPTIONALS`" != 1 ]\
-   && ! grep -Eq 'u:object_r:hal_dms_hwservice:s0|u:object_r:default_android_hwservice:s0' $FILE; then
+  && ! df -h $SYSTEM_EXT | grep 100%\
+  && ! grep -Eq 'u:object_r:hal_dms_hwservice:s0|u:object_r:default_android_hwservice:s0' $FILE; then
     FILE=$SYSTEM_EXT/etc/selinux/system_ext_hwservice_contexts
     patch_hwservice
   fi
@@ -674,11 +696,6 @@ if [ $DOLBY == true ]; then
       ui_print "! Failed to set hal_dms_hwservice context."
       ui_print " "
     fi
-    FILES="$MAGISKTMP/mirror/*/etc/selinux/*_hwservice_contexts
-           $MAGISKTMP/mirror/*/*/etc/selinux/*_hwservice_contexts
-           /*/etc/selinux/*_hwservice_contexts
-           /*/*/etc/selinux/*_hwservice_contexts"
-    restore
   fi
 fi
 
@@ -850,12 +867,13 @@ fi
 dolby_settings() {
 FILE=$MODPATH/system/vendor/etc/dolby/dax-default.xml
 PROP=`grep_prop dolby.bass $OPTIONALS`
-if [ "$PROP" == def ]; then
-  ui_print "- Using default settings of bass-enhancer"
-elif [ "$PROP" == true ]; then
+if [ "$PROP" == true ]; then
   ui_print "- Changing all bass-enhancer-enable value to true"
   sed -i 's|bass-enhancer-enable value="false"|bass-enhancer-enable value="true"|g' $FILE
-elif [ "$PROP" ] && [ "$PROP" != false ] && [ "$PROP" -gt 0 ]; then
+elif [ "$PROP" == false ]; then
+  ui_print "- Changing all bass-enhancer-enable value to false"
+  sed -i 's|bass-enhancer-enable value="true"|bass-enhancer-enable value="false"|g' $FILE
+elif [ "$PROP" ] && [ "$PROP" != def ] && [ "$PROP" -gt 0 ]; then
   ui_print "- Changing all bass-enhancer-enable value to true"
   sed -i 's|bass-enhancer-enable value="false"|bass-enhancer-enable value="true"|g' $FILE
   ROWS=`grep bass-enhancer-boost $FILE | sed -e 's|<bass-enhancer-boost value="||g' -e 's|"/>||g'`
@@ -865,9 +883,6 @@ elif [ "$PROP" ] && [ "$PROP" != false ] && [ "$PROP" -gt 0 ]; then
   for ROW in $ROWS; do
     sed -i "s|bass-enhancer-boost value=\"$ROW\"|bass-enhancer-boost value=\"$PROP\"|g" $FILE
   done
-else
-  ui_print "- Changing all bass-enhancer-enable value to false"
-  sed -i 's|bass-enhancer-enable value="true"|bass-enhancer-enable value="false"|g' $FILE
 fi
 if [ "`grep_prop dolby.virtualizer $OPTIONALS`" == 1 ]; then
   ui_print "- Changing all virtualizer-enable value to true"
@@ -928,12 +943,14 @@ fi
 
 # function
 rename_file() {
-ui_print "- Renaming"
-ui_print "$FILE"
-ui_print "  to"
-ui_print "$MODFILE"
-mv -f $FILE $MODFILE
-ui_print " "
+if [ -f $FILE ]; then
+  ui_print "- Renaming"
+  ui_print "$FILE"
+  ui_print "  to"
+  ui_print "$MODFILE"
+  mv -f $FILE $MODFILE
+  ui_print " "
+fi
 }
 change_name() {
 if grep -q $NAME $FILE; then
@@ -948,6 +965,16 @@ fi
 # mod
 if [ $DOLBY == true ]\
 && [ "`grep_prop dolby.mod $OPTIONALS`" != 0 ]; then
+  NAME=dax-default.xml
+  NAME2=dap-default.xml
+  FILE=$MODPATH/system/vendor/etc/dolby/$NAME
+  MODFILE=$MODPATH/system/vendor/etc/dolby/$NAME2
+  rename_file
+  FILE=$MODPATH/system/vendor/lib*/libdlbdsservice.so
+  change_name
+  NAME=dax_sqlite3.db
+  NAME2=dap_sqlite3.db
+  change_name
   NAME=libswdap.so
   NAME2=libswdlb.so
   if [ "$IS64BIT" == true ]; then
@@ -974,6 +1001,28 @@ $MODPATH/.aml.sh"
 $MODPATH/system/vendor/lib*/vendor.dolby*.hardware.dms*@*-impl.so
 $MODPATH/system/vendor/bin/hw/vendor.dolby*.hardware.dms*@*-service"
   change_name
+  sed -i 's|ro.dolby.mod_uuid false|ro.dolby.mod_uuid true|g' $MODPATH/service.sh
+  NAME=$'\x39\x53\x7a\x04\xbc\xaa'
+  NAME2=_ryuki
+  FILE=$MODPATH/system/vendor/lib*/soundfx/libswdlb.so
+  change_name
+  NAME=$'\x45\x27\x99\x21\x85\x39'
+  FILE=$MODPATH/system/vendor/lib*/soundfx/libswdlb.so
+#  change_name
+  NAME=$'\xd5\x3e\x26\xda\x02\x53'
+  FILE=$MODPATH/system/vendor/lib*/soundfx/libhwdlb.so
+#  change_name
+  NAME=$'\xef\x93\x7f\x67\x55\x87'
+  FILE=$MODPATH/system/vendor/lib*/soundfx/lib*wdlb.so
+  change_name
+  NAME=39537a04bcaa
+  NAME2=5f7279756b69
+  FILE=$MODPATH/.aml.sh
+  change_name
+  NAME=452799218539
+#  change_name
+  NAME=d53e26da0253
+#  change_name
 fi
 
 # function
